Using one Instance for AntiBot
